OpenAlex/requirements.txt
# Core dependencies
requests==2.32.4
PyPDF2==3.0.1
pdfplumber==0.11.7

# For OCR
pdf2image==1.17.0
pytesseract==0.3.13
pillow==11.3.0
fitz==0.0.1.dev2

# For summarization
transformers==4.55.2
torch==2.2.2

frontend==0.0.3
